move definition of util functions
